Cleaned up /molecules - restructured and added doc. to new changes
