+++ frontend/src/utils/constants.js
@@ -87,3 +87,10 @@ function parseIceServers() {
 }
 
 export const PARTICIPANT_HOST = "http://localhost:3000/experimentRoom/";
+
+export const instructionsList = [
+  "Please remove any glasses, caps or other such articles if you are wearing any.",
+  "Ensure that your surrounding lighting is good.",
+  "We would like to know about your experience, so please take 5  minutes at the end \
+  to do the Feedback Survey."
+]
